+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * User receives an error when searching for an item in the Inventory app. Fixes UIIN-2634.
 * Create new instance success toast no longer shows the instance HRID. Fixes UIIN-2635.
 * Optimistic locking message not working for instances in non-consortial tenant. Fixes UIIN-2628.
+* Add immediate warning message when a local instance is shared. Refs UIIN-2617.
 
 ## [10.0.0](https://github.com/folio-org/ui-inventory/tree/v10.0.0) (2023-10-13)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v9.4.12...v10.0.0)